deep_painting_app package with data_processing module and explore_data module - first push
